3 reports are ready as pdf output

+++ resources/views/report/partials/api/profit_by_categories.blade.php
@@ -18,9 +18,9 @@
                             {{$record->category}}
                         </td>
                         <td>
-                            {{ round(intval(strip_tags($record->gross_profit)),2) }}
+                            {{ round(floatval(strip_tags($record->gross_profit)),2) }}
                             @php
-                                $total += intval(strip_tags($record->gross_profit));
+                                $total += floatval(strip_tags($record->gross_profit));
                             @endphp
                         </td>
                     </tr>
